Fix: GCC 13 -Werror build failures (format, sign-compare, unused-result)
- ssl_msg.c: replace MBEDTLS_PRINTF_LONGLONG with %llu for unsigned args
- ssl_msg.c: use named local var to suppress warn_unused_result on best-effort CID replenishment call
- ssl_tls13_server.c: cast to size_t before comparison to avoid sign-compare warning

+++ library/ssl_tls13_server.c
@@ -1713,7 +1713,7 @@ static int ssl_tls13_parse_client_hello(mbedtls_ssl_context *ssl,
                         return MBEDTLS_ERR_SSL_DECODE_ERROR;
                     }
                     cookie_data_len = MBEDTLS_GET_UINT16_BE(p, 0);
-                    if (cookie_data_len + 2 != extension_data_len) {
+                    if ((size_t) cookie_data_len + 2 != extension_data_len) {
                         MBEDTLS_SSL_PEND_FATAL_ALERT(
                             MBEDTLS_SSL_ALERT_MSG_DECODE_ERROR,
                             MBEDTLS_ERR_SSL_DECODE_ERROR);
